[SW-1483]  Deprecate algos and features in org.apache.spark package
