+++ benchmarks/src/main/scala/ai/h2o/sparkling/benchmarks/BenchmarkBase.scala
@@ -19,9 +19,8 @@ package ai.h2o.sparkling.benchmarks
 
 import java.io.{OutputStream, PrintWriter}
 
-import ai.h2o.sparkling.ml.algos.H2OAlgorithm
+import ai.h2o.sparkling.ml.algos.{H2OAlgorithm, H2OGBM, H2OGLM}
 import org.apache.spark.h2o.H2OContext
-import org.apache.spark.ml.h2o.algos.{H2OGBM, H2OGLM}
 import org.apache.spark.sql.SparkSession
 
 import scala.collection.mutable.ArrayBuffer
