+++ ml/src/main/scala/org/apache/spark/ml/h2o/algos/H2OGBM.scala
@@ -14,89 +14,13 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
+
 package org.apache.spark.ml.h2o.algos
 
-import ai.h2o.sparkling.ml.algos.H2OSupervisedAlgorithm
-import ai.h2o.sparkling.ml.params.H2OAlgoSharedTreeParams
-import hex.schemas.GBMV3.GBMParametersV3
-import hex.tree.gbm.GBMModel.GBMParameters
-import hex.tree.gbm.{GBM, GBMModel}
-import org.apache.spark.ml.util.{DefaultParamsReadable, Identifiable}
+import ai.h2o.sparkling.ml.utils.H2OParamsReadable
 
-/**
-  * H2O GBM algorithm exposed via Spark ML pipelines.
-  */
-class H2OGBM(override val uid: String) extends H2OSupervisedAlgorithm[GBM, GBMModel, GBMParameters] with H2OGBMParams {
+@Deprecated
+class H2OGBM(override val uid: String) extends ai.h2o.sparkling.ml.algos.H2OGBM(uid)
 
-  def this() = this(Identifiable.randomUID("gbm"))
-}
-
-object H2OGBM extends DefaultParamsReadable[py_sparkling.ml.algos.H2OGBM]
-
-/**
-  * Parameters for Spark's API exposing underlying H2O model.
-  */
-trait H2OGBMParams extends H2OAlgoSharedTreeParams[GBMParameters] {
-
-  type H2O_SCHEMA = GBMParametersV3
-
-  protected def paramTag = reflect.classTag[GBMParameters]
-
-  protected def schemaTag = reflect.classTag[H2O_SCHEMA]
-
-  //
-  // Param definitions
-  //
-  private val learnRate = doubleParam("learnRate")
-  private val learnRateAnnealing = doubleParam("learnRateAnnealing")
-  private val colSampleRate = doubleParam("colSampleRate")
-  private val maxAbsLeafnodePred = doubleParam("maxAbsLeafnodePred")
-  private val predNoiseBandwidth = doubleParam("predNoiseBandwidth")
-
-  //
-  // Default values
-  //
-  setDefault(
-    learnRate -> parameters._learn_rate,
-    learnRateAnnealing -> parameters._learn_rate_annealing,
-    colSampleRate -> parameters._col_sample_rate,
-    maxAbsLeafnodePred -> parameters._max_abs_leafnode_pred,
-    predNoiseBandwidth -> parameters._pred_noise_bandwidth
-  )
-
-  //
-  // Getters
-  //
-  def getLearnRate(): Double = $(learnRate)
-
-  def getLearnRateAnnealing(): Double = $(learnRateAnnealing)
-
-  def getColSampleRate(): Double = $(colSampleRate)
-
-  def getMaxAbsLeafnodePred(): Double = $(maxAbsLeafnodePred)
-
-  def getPredNoiseBandwidth(): Double = $(predNoiseBandwidth)
-
-  //
-  // Setters
-  //
-  def setLearnRate(value: Double): this.type = set(learnRate, value)
-
-  def setLearnRateAnnealing(value: Double): this.type = set(learnRateAnnealing, value)
-
-  def setColSampleRate(value: Double): this.type = set(colSampleRate, value)
-
-  def setMaxAbsLeafnodePred(value: Double): this.type = set(maxAbsLeafnodePred, value)
-
-  def setPredNoiseBandwidth(value: Double): this.type = set(predNoiseBandwidth, value)
-
-
-  override def updateH2OParams(): Unit = {
-    super.updateH2OParams()
-    parameters._learn_rate = $(learnRate)
-    parameters._learn_rate_annealing = $(learnRateAnnealing)
-    parameters._col_sample_rate = $(colSampleRate)
-    parameters._max_abs_leafnode_pred = $(maxAbsLeafnodePred)
-    parameters._pred_noise_bandwidth = $(predNoiseBandwidth)
-  }
-}
+@Deprecated
+object H2OGBM extends H2OParamsReadable[ai.h2o.sparkling.ml.algos.H2OGBM]
