+++ py/ai/h2o/sparkling/ml/features/__init__.py
@@ -16,3 +16,4 @@
 #
 
 from ai.h2o.sparkling.ml.features.H2OTargetEncoder import H2OTargetEncoder
+from ai.h2o.sparkling.ml.features.ColumnPruner import ColumnPruner
